+++ doc/conf.py
@@ -57,7 +57,7 @@ copyright = u'2017-2020, Ken Kundert'
 # built documents.
 #
 # The full version, including alpha/beta/rc tags.
-release = '2.10.4'
+release = '2.10.5'
 # The short X.Y version.
 version = '.'.join(release.split('.')[0:2])