+++ quantiphy.py
@@ -673,8 +673,8 @@ def add_constant(value, alias=None, unit_systems=None):
 
 
 # Globals {{{1
-__version__ = '2.10.4'
-__released__ = '2020-07-01'
+__version__ = '2.10.5'
+__released__ = '2020-07-18'
 
 # These mappings are only used when reading numbers
 # The key for these mappings must be a single character